[Refactor] TabBar에 Navigation Controller가 다시 사용되는 것을 고려하여 Appearance를 설정하는 메서드를 BaseNavigationController를 만들어 재적용
